requirements38.txt
asgiref==3.2.7
attrs==19.3.0
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
configparser==5.0.1
crayons==0.4.0
decorator==4.4.2
Django==3.0.6
djangorestframework==3.12.2
idna==2.10
importlib-metadata==1.6.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
more-itertools==8.3.0
mysqlclient==1.4.6
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.6
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
pytest-pythonpath==0.7.3
pytz==2020.1
requests==2.25.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webdriver-manager==3.2.2
zipp==3.1.0
